timebase.c: Fix GPSD time will jump back 1024 weeks
This patch fixes the issue reported here:
https://gitlab.com/gpsd/gpsd/-/issues/144

The time would jump back to 2002 after October 23, 2021.  This should
now be resolved in the 3.20 released version of gpsd.

The issue was originally introduced here:
https://gitlab.com/gpsd/gpsd/-/commit/cb7367496f6dd5f90c6b067c031b327921a857b2

The issue was resolved for the 3.23 release here:
https://gitlab.com/gpsd/gpsd/-/commit/7f30d88d04dc62b8bd6265ad1d09d72d220f97f6

This fix is just resolving the jump back of 1024 weeks from those
commits.